packages/contacts-service/requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
click==8.1.8
confluent-kafka==2.7.0
conprotos @ file:///Users/divyam/Documents/projects/conmail/packages/protos-service/dist/conprotos-0.0.1-py3-none-any.whl
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-Pydantic==0.12.0
Flask-SQLAlchemy==3.1.1
googleapis-common-protos==1.66.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.8
MarkupSafe==3.0.2
protobuf==5.29.2
psycopg2==2.9.10
pydantic==2.10.4
pydantic_core==2.27.2
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3